add floating point timer and end address to bulk init

+++ avr/usbload/timer.h
@@ -1,5 +1,5 @@
 
 
 uint16_t timer_start( void );
-uint16_t timer_stop( void );
+double timer_stop( void );
